+++ Changelog.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `VK_EXT_hdr_metadata` device extension (#804)
 - Added `VK_NV_cuda_kernel_launch` device extension (#805)
 - Added `descriptor_count()` setter on `ash::vk::WriteDescriptorSet` (#809)
+- Added `*_as_c_str()` getters for `c_char` pointers and `c_char` arrays (#831)
 
 ### Changed
 
@@ -48,7 +49,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `VK_KHR_device_group_creation`: Take borrow of `Entry` in `fn new()` (#753)
 - `VK_KHR_device_group_creation`: Rename `vk::Instance`-returning function from `device()` to `instance()` (#759)
 - Windows `HANDLE` types (`HWND`, `HINSTANCE`, `HMONITOR`) are now defined as `isize` instead of `*const c_void` (#797)
-- extensions/ext/ray_tracing_pipeline: Pass indirect SBT regions as single item reference. (#829)
+- extensions/ext/ray_tracing_pipeline: Pass indirect SBT regions as single item reference (#829)
+- Replaced `c_char` array setters with `CStr` setters (#831)
 
 ### Removed